requirements.txt
crewai>=0.28.8
crewai-tools>=0.1.6
python-dotenv>=1.0.0